requirements.txt
aiogram==2.25.2
PyPDF2==3.0.1
Pillow==10.2.0
ebooklib==0.18
